Fix paths after UI tests moves.
Add missing layout __tests__ files.
