+++ frontend/src/ui/layout/__tests__/Header.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { configure, mount } from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
-import Header from '../ui/layout/Header'
+import Header from '../Header'
 
 configure({ adapter: new Adapter() })
