+++ frontend/src/ui/layout/__tests__/Logo.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { configure, mount } from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
-import Logo from '../ui/layout/Logo'
+import Logo from '../Logo'
 
 configure({ adapter: new Adapter() })
