+++ frontend/src/ui/layout/__tests__/LogoContainer.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { configure, mount } from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
-import LogoContainer from '../ui/layout/LogoContainer'
+import LogoContainer from '../LogoContainer'
 
 configure({ adapter: new Adapter() })
