+++ frontend/src/ui/presentational/__tests__/BankAccountsList.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { configure, mount } from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
-import BankAccountsList from '../ui/presentational/BankAccountsList'
+import BankAccountsList from '../BankAccountsList'
 
 configure({ adapter: new Adapter() })
