+++ frontend/src/ui/presentational/__tests__/Loading.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { configure, mount } from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
-import Loading from '../ui/presentational/Loading'
+import Loading from '../Loading'
 
 configure({ adapter: new Adapter() })
