+++ frontend/src/ui/presentational/__tests__/VerifiedBankAccountsList.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { configure, mount } from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
-import VerifiedBankAccountsList from '../ui/presentational/VerifiedBankAccountsList'
+import VerifiedBankAccountsList from '../VerifiedBankAccountsList'
 
 configure({ adapter: new Adapter() })
